requirements.txt
openai
click
pyyaml 